backend/requirements.txt
flask>=2.0
flask-cors>=3.0
werkzeug>=2.0
pytest>=7.0
streamlit>=1.0
numpy>=1.21
Pillow>=8.0
requests>=2.0

# Backend / DB / ML
tensorflow>=2.0
pymongo>=4.0
python-dotenv>=0.21.0
tensorflow-cpu>=2.5
pymongo>=3.11